Fix ArticlePage folder structure
