+++ apps/codeforafrica/src/components/ArticlePage/ArticlePage.js
@@ -1,0 +1,100 @@
+import { RichTypography, Section } from "@commons-ui/core";
+import { Box, Typography, Divider } from "@mui/material";
+import Image from "next/image";
+import PropTypes from "prop-types";
+import React from "react";
+
+import ArticleHeader from "@/codeforafrica/components/ArticleHeader";
+import Author from "@/codeforafrica/components/Author";
+import ShareBar from "@/codeforafrica/components/ShareBar";
+import {
+  FacebookShareBarButton,
+  LinkedinShareBarButton,
+  TwitterShareBarButton,
+} from "@/codeforafrica/components/ShareBarButton";
+
+const ArticlePage = React.forwardRef(function ArticlePage(props, ref) {
+  const {
+    coverImage,
+    title,
+    date,
+    tags,
+    summary,
+    author,
+    profession,
+    content,
+    ...other
+  } = props;
+
+  return (
+    <Box ref={ref} {...other}>
+      <Box
+        component="figure"
+        sx={{
+          position: "relative",
+          width: "100%",
+          margin: 0,
+          height: { xs: "163px", md: "600px" },
+        }}
+      >
+        <Image
+          alt={title}
+          src={coverImage?.src}
+          layout="fill"
+          objectFit="cover"
+        />
+      </Box>
+      <Section
+        sx={{
+          px: { xs: 2.5, sm: 0 },
+          maxWidth: {
+            sm: "648px",
+            md: "912px",
+          },
+        }}
+        ref={ref}
+        {...other}
+      >
+        <ArticleHeader
+          title={title}
+          date={date}
+          tags={tags}
+          summary={summary}
+        />
+        <Box
+          sx={{
+            color: "grey.main",
+            columnGap: 2,
+            display: "flex",
+            alignItems: "center",
+            py: "20px",
+          }}
+        >
+          <Typography variant="footerCap">Share This Article</Typography>
+          <ShareBar>
+            <FacebookShareBarButton />
+            <LinkedinShareBarButton />
+            <TwitterShareBarButton />
+          </ShareBar>
+        </Box>
+        <RichTypography variant="body3" sx={{ my: { xs: 2.5, md: 5 } }}>
+          {content}
+        </RichTypography>
+        <Divider />
+        <Author author={author} profession={profession} />
+      </Section>
+    </Box>
+  );
+});
+
+ArticlePage.propTypes = {
+  content: PropTypes.string,
+  article: PropTypes.shape({}),
+};
+
+ArticlePage.defaultProps = {
+  content: undefined,
+  article: undefined,
+};
+
+export default ArticlePage;
