+++ apps/codeforafrica/src/components/ArticlePage/ArticlePage.snap.js
@@ -1,0 +1,100 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`<ArticlePage /> renders unchanged 1`] = `
+<div>
+  <div
+    class="MuiBox-root css-0"
+  >
+    <figure
+      class="MuiBox-root css-1i6lsvp"
+    >
+      <span
+        style="box-sizing: border-box; display: block; overflow: hidden; background: none; opacity: 1; border: 0px; margin: 0px; padding: 0px; position: absolute; top: 0px; left: 0px; bottom: 0px; right: 0px;"
+      >
+        <img
+          data-nimg="fill"
+          decoding="async"
+          src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7"
+          style="position: absolute; top: 0px; left: 0px; bottom: 0px; right: 0px; box-sizing: border-box; padding: 0px; margin: auto; display: block; width: 0px; height: 0px; min-width: 100%; max-width: 100%; min-height: 100%; max-height: 100%; object-fit: cover;"
+        />
+        <noscript />
+      </span>
+    </figure>
+    <div
+      class="MuiContainer-root MuiContainer-maxWidthLg MuiContainer-fixed MuiContainer-disableGutters css-1hx1xv5-MuiContainer-root"
+    >
+      <div
+        class="MuiBox-root css-j7qwjs"
+      />
+      <div
+        class="MuiBox-root css-1bswtiy"
+      >
+        <span
+          class="MuiTypography-root MuiTypography-footerCap css-3358n1-MuiTypography-root"
+        >
+          Share This Article
+        </span>
+        <div
+          class="css-k5dji9-ShareBar-root"
+        >
+          <button
+            aria-label="Facebook"
+            class="react-share__ShareButton"
+            data-mui-internal-clone-element="true"
+            style="background-color: transparent; padding: 0px; cursor: pointer;"
+          >
+            <div
+              aria-hidden="true"
+              class="MuiSvgIcon-root MuiSvgIcon-fontSizeMedium css-i4bv87-MuiSvgIcon-root"
+              focusable="false"
+              viewbox="0 0 24 24"
+            />
+          </button>
+          <button
+            aria-label="Linkedin"
+            class="react-share__ShareButton"
+            data-mui-internal-clone-element="true"
+            style="background-color: transparent; padding: 0px; cursor: pointer;"
+          >
+            <div
+              aria-hidden="true"
+              class="MuiSvgIcon-root MuiSvgIcon-fontSizeMedium css-i4bv87-MuiSvgIcon-root"
+              focusable="false"
+              viewbox="0 0 24 24"
+            />
+          </button>
+          <button
+            aria-label="Twitter"
+            class="react-share__ShareButton"
+            data-mui-internal-clone-element="true"
+            style="background-color: transparent; padding: 0px; cursor: pointer;"
+          >
+            <div
+              aria-hidden="true"
+              class="MuiSvgIcon-root MuiSvgIcon-fontSizeMedium css-i4bv87-MuiSvgIcon-root"
+              focusable="false"
+              viewbox="0 0 24 24"
+            />
+          </button>
+        </div>
+      </div>
+      <hr
+        class="MuiDivider-root MuiDivider-fullWidth css-9mgopn-MuiDivider-root"
+      />
+      <div
+        class="MuiBox-root css-3vtd9l"
+      >
+        <span
+          class="MuiTypography-root MuiTypography-body2 css-16etphd-MuiTypography-root"
+        >
+          Article by
+           
+        </span>
+        <div
+          class="MuiBox-root css-hjoi4d"
+        />
+      </div>
+    </div>
+  </div>
+</div>
+`;
