+++ apps/codeforafrica/src/components/ArticlePage/ArticlePage.test.js
@@ -1,16 +1,20 @@
 import { createRender } from "@commons-ui/testing-library";
 import React from "react";
 
-import ArticlePage from ".";
+import ArticlePage from "./ArticlePage";
 
 import theme from "@/codeforafrica/theme";
 
 // eslint-disable-next-line testing-library/render-result-naming-convention
 const render = createRender({ theme });
 
+const defaultProps = {
+  coverImage: { src: "/images/stories-1.png" },
+};
+
 describe("<ArticlePage />", () => {
   it("renders unchanged", () => {
-    const { container } = render(<ArticlePage />);
+    const { container } = render(<ArticlePage {...defaultProps} />);
     expect(container).toMatchSnapshot();
   });
 });
