--- apps/codeforafrica/src/components/ArticlePage/index.snap.js
@@ -1,100 +1,0 @@
-// Jest Snapshot v1, https://goo.gl/fbAQLP
-
-exports[`<ArticlePage /> renders unchanged 1`] = `
-<div>
-  <div
-    class="MuiContainer-root MuiContainer-maxWidthLg MuiContainer-fixed MuiContainer-disableGutters css-usii6f-MuiContainer-root"
-  >
-    <div
-      class="MuiGrid-root MuiGrid-container css-11lq3yg-MuiGrid-root"
-    >
-      <div
-        class="MuiGrid-root MuiGrid-item MuiGrid-grid-xs-12 css-18qart5-MuiGrid-root"
-      >
-        <div
-          class="MuiGrid-root MuiGrid-container css-11lq3yg-MuiGrid-root"
-        >
-          <div
-            class="MuiGrid-root MuiGrid-item MuiGrid-grid-xs-12 css-18qart5-MuiGrid-root"
-          />
-        </div>
-        <div
-          class="MuiBox-root css-1bswtiy"
-        >
-          <span
-            class="MuiTypography-root MuiTypography-footerCap css-3358n1-MuiTypography-root"
-          >
-            Share This Article
-          </span>
-          <div
-            class="css-k5dji9-ShareBar-root"
-          >
-            <button
-              aria-label="Facebook"
-              class="react-share__ShareButton"
-              data-mui-internal-clone-element="true"
-              style="background-color: transparent; padding: 0px; cursor: pointer;"
-            >
-              <svg
-                aria-hidden="true"
-                class="MuiSvgIcon-root MuiSvgIcon-fontSizeMedium css-i4bv87-MuiSvgIcon-root"
-                focusable="false"
-                viewBox="0 0 24 24"
-              />
-            </button>
-            <button
-              aria-label="Linkedin"
-              class="react-share__ShareButton"
-              data-mui-internal-clone-element="true"
-              style="background-color: transparent; padding: 0px; cursor: pointer;"
-            >
-              <svg
-                aria-hidden="true"
-                class="MuiSvgIcon-root MuiSvgIcon-fontSizeMedium css-i4bv87-MuiSvgIcon-root"
-                focusable="false"
-                viewBox="0 0 24 24"
-              />
-            </button>
-            <button
-              aria-label="Twitter"
-              class="react-share__ShareButton"
-              data-mui-internal-clone-element="true"
-              style="background-color: transparent; padding: 0px; cursor: pointer;"
-            >
-              <svg
-                aria-hidden="true"
-                class="MuiSvgIcon-root MuiSvgIcon-fontSizeMedium css-i4bv87-MuiSvgIcon-root"
-                focusable="false"
-                viewBox="0 0 24 24"
-              />
-            </button>
-          </div>
-        </div>
-        <hr
-          class="MuiDivider-root MuiDivider-fullWidth css-9mgopn-MuiDivider-root"
-        />
-        <div
-          class="MuiBox-root css-d944bl"
-        >
-          <span
-            class="MuiTypography-root MuiTypography-body2 css-lupkoz-MuiTypography-root"
-          >
-            Article by
-             
-          </span>
-          <div
-            class="MuiBox-root css-139x83s"
-          >
-            <span
-              class="MuiTypography-root MuiTypography-body1SemiBold css-1s0xknt-MuiTypography-root"
-            />
-            <span
-              class="MuiTypography-root MuiTypography-body1 css-4tj74i-MuiTypography-root"
-            />
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-`;
